tests/iauthd-c.conf: Work around apparent musl or Alpine bug
"make check" would fail on Alpine 3.10.2 with "missing test plan", which
was caused by dlopen() not loading test_sar.so because it had some
unresolved symbols (such as test_is) that should be provided by tests.so.

There was not an obvious reason that dlopen() would effectively ignore
the RTLD_LAZY flag, so use this as a workaround.